편지 보관함 더미 데이터 구현

+++ app/src/main/java/com/example/mailpost/model/Comment.kt
@@ -4,5 +4,5 @@ data class Comment(
     val id: Int,
     val letterId: Int,
     val content: String,
-    val timeInMills: Double,
+    val timeInMills: Long,
 )
